Move xml_files_in back to gen_constants_rs.py
